Fixed some errors

+++ include/World.hpp
@@ -61,7 +61,7 @@ private:
     /// @param x as base-0
     /// @param y as base-0
     /// @return (x, y) as a 1D index
-    uint64_t convert_to_1D_(Vector2 vec) const noexcept;
+    size_t convert_to_1D_(Vector2 vec) const noexcept;
 
     /// @brief checks whether (x, y) is a valid position
     /// @param x as base-0
